extracted ColorDouble
